PowerManager: removed suspend/hibernate features
These features no longer work on recent systems and they are no longer supported with recent versions of UPower (0.99+).
We should support SystemD but suspend/hibernate features are already managed by Logout applet (and the system when pressing the Shutdown button) and this applet can be hidden most of the time (or not used when there are no battery): I think users use Logout applet instead. (and now they should use it :-) )
But note that a revert is always possible ;-)