Recent-Events: create our submenu earlier
On Wayland, it is not easily possible to resize the menu after its creation (i.e. gtk_menu_reposition () seems to do nothing), so adding a menu item after its creation will result in it being "scrolled", making context menus difficult to navigate. Work around this by adding the submenu when the context menu is created and only populating it later when the reply from Zeitgeist is received.