gnome-integration: adapt to new FM API and also use the DBus interface
We could potentially request the logout applet to show confirmation dialogs. We currently do not do this, altough this could be required.

Also, instead of running commands, we can directly use the org.gnome.SessionManager DBus interface. We also check the related GSetting to determine whether to show the confirmation prompt.