Status-Notifier: workaround incorrect object paths supplied by ayatana
ayatana-indicator-application.service seems to assume that the menu path is an addition of the indicator object path and reports the menu path instead of the object path (expecting us to correct for it? or assuming we do not need to communicate with the StatusNotifierItem interface at all?). Workaround this by assuming the default object path if the supplied path obviously corresponds to a menu.