DBus: disable the interface while the dock is reloading
Since this applet does not have a stop method, the main interface will stay available during a reload (even if the module instance is deleted). In theory, this could lead to crashes if we access non-existent objects (including the config). Not sure if this can really happen, since reloading the dock is a synchronous operation and thus we likely do not get any incoming events during it, but better to explicitly avoid it (e.g. gtk-layer-shell does roundtrips in some cases, GTK may decide to iterate the main loop, etc.).